Changed x and z to double for more accurate coords

+++ src/com/robinmc/ublisk/utils/NPCUtils.java
@@ -34,7 +34,7 @@ public class NPCUtils {
 	 * @param name Name of NPC as will appear floating above it
 	 * @param skin In-game name of the skin this NPC will have
 	 */
-	public void spawnNPC(int x, int y, int z, int pitch, int yaw, String name, String skin){
+	public void spawnNPC(double x, int y, double z, int pitch, int yaw, String name, String skin){
 		Location loc = new Location(Var.world(), x, y, z, pitch, yaw);
 		NPCLib.spawnPlayerNPC(loc, name, skin);
 	}
